% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedStat.R
\name{bankStat}
\alias{bankStat}
\title{Banked Statistic}
\usage{
bankStat(bankObj, testSet, bankMatrix)
}
\arguments{
\item{bankObj}{Compressed representation of the cohort being compared to.}

\item{testSet}{Cohort representing the bank.}

\item{bankMatrix}{The actual bank.}
}
\value{
Matrix of p values.
}
\description{
Banked Newman Statistic for comparing individual samples to a cohort of similar samples.
}
\examples{
input1 <- read.csv(file = "filepath.csv",row.names = 1) #Bank Object consisting of row means, estimate of std deviation and background distribution
input2 <- read.csv(file = "filepath.csv", row.names = 1) #Dataset to be tested against the bank
input3 <- read.csv(file = "filepath.csv", row.names = 1) #Dataset that comprises the bank

RN <- rownames(input2)

bankObj <- data.matrix(input1)
testSet <- data.matrix(input2)
bankMatrix <- data.matrix(input3)

bankStat(bankObj, testSet, bankMatrix)
}
